Migration to JME 3.8.1(LWJGL3), ShadingType added, NativeCanvas deprecated, 'platform-homebrew' migration incomplete

+++ platform-jme/src/main/java/de/yard/threed/platform/jme/JmeCamera.java
@@ -412,7 +412,7 @@ public class JmeCamera implements NativeCamera/*, NativeTransform */ {
 
     @Override
     public void setFar(double far) {
-        camera.setFrustumFar(far);
+        camera.setFrustumFar( (float)far);
     }
 
     @Override
